% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_formulas}
\alias{check_formulas}
\title{check_formulas}
\usage{
check_formulas(
  measurement,
  structure,
  allowed = "^[a-zA-Z0-9_]+[ ]*~[a-zA-Z0-9_\\\\+ ]+$"
)
}
\arguments{
\item{measurement}{alist with the measurement formulas}

\item{structure}{alist with the structural formulas}

\item{allowed}{pattern of supported formulas}
}
\value{
nothing. Throws error in case of unsupported formulas
}
\description{
R's formulas are much more flexible than the current framework of plsR
supports. We have to make sure that users are not trying to fit more complex
models than what plsR allows.
}
\examples{
plsR:::check_formulas(measurement = alist(C ~ x1 + x2,
                                              D ~ x3 + x4),
                          structure = alist(D ~ C))

}
