% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{coef.PLS_SEM}
\alias{coef.PLS_SEM}
\title{coef.PLS_SEM}
\usage{
\method{coef}{PLS_SEM}(object, ...)
}
\arguments{
\item{object}{PLS_SEM object}

\item{...}{not used}
}
\value{
vector with parameters
}
\description{
Coefficient of PLS SEM
}
