% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSP.R
\name{coef.PLSP_SEM}
\alias{coef.PLSP_SEM}
\title{coef.PLSP_SEM}
\usage{
\method{coef}{PLSP_SEM}(object, ...)
}
\arguments{
\item{object}{PLSP_SEM object}

\item{...}{not used}
}
\value{
vector with parameters
}
\description{
Coefficient of PLSP SEM
}
