% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{compute_summarized_effects}
\alias{compute_summarized_effects}
\title{compute_summarized_effects}
\usage{
compute_summarized_effects(PLS_result)
}
\arguments{
\item{PLS_result}{PLS model fitted with PLS() (see ?PLS)}
}
\value{
list with direct, indirect, and total effects.
}
\description{
Computes direct, indirect, and total effects
for a PLS model
}
\keyword{internal}
