% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{confidence_intervals}
\alias{confidence_intervals}
\title{confidence_intervals}
\usage{
confidence_intervals(
  PLS_result,
  include_loadings = FALSE,
  alpha = 0.05,
  R = 1000
)
}
\arguments{
\item{PLS_result}{results from PLS}

\item{include_loadings}{should loadings also be returned?}

\item{alpha}{significance level}

\item{R}{number of repetitions}
}
\value{
boostrap results
}
\description{
Compute confidence intervals for the parameters using bootstrap samples
}
