% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent_PLS.R
\name{debias_correlation}
\alias{debias_correlation}
\title{debias_correlation}
\usage{
debias_correlation(PLS_result)
}
\arguments{
\item{PLS_result}{model estimated with PLS()}
}
\value{
matrix with debiased correlations for reflective constructs
}
\description{
Debias the correlations of reflective constructs
See Dijkstra, T. K., & Henseler, J. (2015). Consistent partial least squares path modeling. MIS quarterly, 39(2), 297-316.
}
