% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent_PLS.R
\name{debias_PLS}
\alias{debias_PLS}
\title{debias_PLS}
\usage{
debias_PLS(PLS_result)
}
\arguments{
\item{PLS_result}{a model fitted with PLS()}
}
\value{
the model with updated loadings and regressions
}
\description{
Debiases the regressions and loadings
in a PLS model with reflective constructs.
See Dijkstra, T. K., & Henseler, J. (2015). Consistent partial least squares path modeling. MIS quarterly, 39(2), 297-316.
}
\keyword{internal}
