% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_treatment.R
\name{fail_on_NA}
\alias{fail_on_NA}
\title{fail_on_NA}
\usage{
fail_on_NA(data, weights)
}
\arguments{
\item{data}{data set with missings}

\item{weights}{vector with weights for each person in the data set}
}
\value{
data in case of no error
}
\description{
Fails if there is any missing data
}
\examples{
library(basicPLS)
satisfaction_with_missings <- basicPLS::satisfaction
missings <- matrix(sample(c(TRUE, FALSE),
                   nrow(basicPLS::satisfaction)*ncol(basicPLS::satisfaction),
                   prob = c(.1, .9),
                   replace = TRUE),
                   nrow = nrow(basicPLS::satisfaction),
                   ncol = ncol(basicPLS::satisfaction))
satisfaction_with_missings[missings] <- NA
try(fail_on_NA(data = satisfaction_with_missings,
               weights = rep(1, nrow(satisfaction_with_missings))))
}
