% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{flatten_effects}
\alias{flatten_effects}
\title{flatten_effects}
\usage{
flatten_effects(effects, separator = "<-")
}
\arguments{
\item{effects}{summarized total or indirect effects}

\item{separator}{separator used when renaming the parameters}
}
\value{
vector with effects
}
\description{
Given a list of direct or indirect effects, this function flattens the list
to a vector with relabeled parameters
}
\keyword{internal}
