% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent_PLS.R
\name{get_reliability}
\alias{get_reliability}
\title{get_reliability}
\usage{
get_reliability(PLS_result)
}
\arguments{
\item{PLS_result}{model estimated with PLS()}
}
\value{
vector with reliabilities. For composites, the reliability is 1.0
}
\description{
Compute the reliabilities of formative constructs in a PLS model
}
\examples{
library(basicPLS)
data_set <- basicPLS::satisfaction

# Both, measurement and structural model are specified using R's formulas:
PLS_result <- PLS(
  measurement = alist(EXPE ~ expe1 + expe2 + expe3 + expe4 + expe5,
                      IMAG ~ imag1 + imag2 + imag3 + imag4 + imag5,
                      LOY ~ loy1 + loy2 + loy3 + loy4,
                      QUAL ~ qual1 + qual2 + qual3 + qual4 + qual5,
                      SAT ~ sat1 + sat2 + sat3 + sat4,
                      VAL ~ val1 + val2 + val3 + val4),
  structure = alist(QUAL ~ EXPE,
                    EXPE ~ IMAG,
                    SAT ~ IMAG + EXPE + QUAL + VAL,
                    LOY ~ IMAG + SAT,
                    VAL ~ EXPE + QUAL),
  as_reflective = c("EXPE", "IMAG", "LOY", "QUAL", "SAT", "VAL"),
  data = data_set)
PLS_result

get_reliability(PLS_result)
}
