% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{mode_A}
\alias{mode_A}
\title{mode_A}
\usage{
mode_A(measurement, composite, data_std, composite_values, sample_weights)
}
\arguments{
\item{measurement}{formula specifying the measurement (e.g., C ~ c_1 + c_2 + c_3)}

\item{composite}{the name of the composite which should be computed as (e.g., "C")}

\item{data_std}{standardized data set}

\item{composite_values}{current composite values}

\item{sample_weights}{weight for each observation}
}
\value{
a vector with updated weights
}
\description{
Mode A computes the weights as pseudo loadings. Say, composite C is a combination
of c_1, c_2, c_3 (e.g., C = .4*c_1 + .6*c_2 + .2*c_3). Then, mode A turns the
regression around and predicts each of the items c_1, c_2, c_3 with composite C:
c_1 = b_01 + b_11 * C + e, c_2 = b_02 + b_12 * C + e, c_3 = b_03 + b_13 * C + e.
The new weights are given by b_11, b_12, b_13.
}
