% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{mode_B}
\alias{mode_B}
\title{mode_B}
\usage{
mode_B(measurement, composite, data_std, composite_values, sample_weights)
}
\arguments{
\item{measurement}{formula specifying the measurement (e.g., C ~ c_1 + c_2 + c_3)}

\item{composite}{the name of the composite which should be computed as (e.g., "C")}

\item{data_std}{standardized data set}

\item{composite_values}{current composite values}

\item{sample_weights}{weight for each observation}
}
\value{
a vector with updated weights
}
\description{
Mode B computes the weights using a regression of the composite on the items.
Say, composite C is a combination of c_1, c_2, c_3 (e.g., C = .4*c_1 + .6*c_2 + .2*c_3).
Then, mode B estimates the new weights by estimating the weights of the regression
C = b_0 + b_1*c_1 + b_2*c_2 + b_3*c_3 + e.
The new weights are given by b_1, b_2, b_3.
}
