% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{PLS}
\alias{PLS}
\title{PLS}
\usage{
PLS(measurement, structure, data, max_iterations = 1000, convergence = 1e-05)
}
\arguments{
\item{measurement}{alist with formulas defining the measurements}

\item{structure}{alist with formulas defining structural model}

\item{data}{data set (data.frame)}

\item{max_iterations}{maximal number of iterations for the estimation
algorithm}

\item{convergence}{convergence criterion. If the maximal change in the
weights falls below this value, the estimation is stopped.}
}
\value{
list with weights, effects, and components
}
\description{
Basic function estimating a PLS-SEM with formative measurement model.
The algorithm is adapted from https://github.com/gastonstat/plspm.
}
\examples{
library(cSEM)
data_set <- 10*cSEM::threecommonfactors + 2
PLS_result <- PLS(measurement = alist(eta1 ~ y11 + y12 + y13,
                                      eta2 ~ y21 + y22 + y23,
                                      eta3 ~ y31 + y32 + y33),
                  structure = alist(eta2 ~ eta1,
                                    eta3 ~ eta1 + eta2),
                  data = data_set)

# same thing with cSEM
model <- "
# Structural model
eta2 ~ eta1
eta3 ~ eta1 + eta2

# measurement model
eta1 <~ y11 + y12 + y13
eta2 <~ y21 + y22 + y23
eta3 <~ y31 + y32 + y33
"
fit_csem <- csem(data_set,
                 model,
                 .PLS_modes = "modeB")
# Let's compare the results:
fit_csem$Estimates$Weight_estimates
PLS_result$weights

fit_csem$Estimates$Path_estimates
PLS_result$effects

head(fit_csem$Estimates$Construct_scores -
       PLS_result$components)

assess(fit_csem, "r2")
get_r2(PLS_result)
}
