% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{print.PLS_SEM}
\alias{print.PLS_SEM}
\title{print.PLS_SEM}
\usage{
\method{print}{PLS_SEM}(x, ...)
}
\arguments{
\item{x}{PLS SEM model}

\item{...}{not used}
}
\value{
nothing
}
\description{
Print results of PLS SEM
}
