% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSP.R
\name{print.PLSP_SEM}
\alias{print.PLSP_SEM}
\title{print.PLSP_SEM}
\usage{
\method{print}{PLSP_SEM}(x, ...)
}
\arguments{
\item{x}{PLSP SEM model}

\item{...}{not used}
}
\value{
nothing
}
\description{
Print results of PLSP SEM
}
