% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{regression_coef}
\alias{regression_coef}
\title{regression_coef}
\usage{
regression_coef(formula, data, wt = NULL)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data set}

\item{wt}{sample weights vector}
}
\value{
regression coefficients
}
\description{
Computes regression coefficients and allows weighting. This function is necessary
to re-assign the environment of the formula and make sure that R finds the weights.
}
\keyword{internal}
