% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{regression_from_correlation}
\alias{regression_from_correlation}
\title{For debiasing, we need to compute regressions based on correlations. This
function provides that feature. See https://wviechtb.github.io/metafor/reference/matreg.html}
\usage{
regression_from_correlation(R, dependent, predictors)
}
\arguments{
\item{R}{correlation matrix}

\item{dependent}{name of the dependent variable}

\item{predictors}{vector with names of the predictors}
}
\value{
regression coefficients
}
\description{
For debiasing, we need to compute regressions based on correlations. This
function provides that feature. See https://wviechtb.github.io/metafor/reference/matreg.html
}
\keyword{internal}
