% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{satisfaction}
\alias{satisfaction}
\title{satisfaction}
\format{
An object of class \code{data.frame} with 250 rows and 29 columns.
}
\usage{
satisfaction
}
\description{
Satisfaction data copied from cSEM. The data has one additional column: A vector
with sample weights.
See ?cSEM::satisfaction for more details.
}
\references{
The data set is copied from the cSEM package.
}
\keyword{data}
