% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{satisfaction_NA}
\alias{satisfaction_NA}
\title{satisfaction_NA}
\format{
An object of class \code{data.frame} with 250 rows and 29 columns.
}
\usage{
satisfaction_NA
}
\description{
Satisfaction data copied from cSEM. The data has one additional column: A vector
with sample weights. Additionally, about 10% of the observations are missing.
See ?cSEM::satisfaction for more details.
}
\references{
The data set is copied from the cSEM package.
}
\keyword{data}
