% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{standard_errors}
\alias{standard_errors}
\title{standard_errors}
\usage{
standard_errors(PLS_result, R = 1000)
}
\arguments{
\item{PLS_result}{results from PLS}

\item{R}{number of repetitions}
}
\value{
boostrap results
}
\description{
Compute standard errors using bootstrap samples
}
