% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLS.R
\name{confidence_intervals}
\alias{confidence_intervals}
\title{confidence_intervals}
\usage{
confidence_intervals(PLS_result, alpha = 0.05, R = 1000, ...)
}
\arguments{
\item{PLS_result}{results from PLS}

\item{alpha}{significance level}

\item{R}{number of repetitions}

\item{...}{additional arguments passed to boot. See ?boot::boot}
}
\value{
bootstrap results
}
\description{
Compute confidence intervals for the parameters using bootstrap samples
}
